requirements.txt
absl-py==2.3.1
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.55.3
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
mediapipe==0.10.21
ml_dtypes==0.5.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
packaging==24.2
pandas==2.2.3
parso==0.8.4
phonenumbers==9.0.7
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==4.25.8
pure_eval==0.2.3
pycparser==2.22
pyfiglet==1.0.3
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
setuptools==72.1.0
six==1.17.0
sortedcontainers==2.4.0
sounddevice==0.5.2
sqlparse==0.5.0
stack-data==0.6.3
threadpoolctl==3.6.0
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.44.0
